update to all pages
