+++ app/companies/page.tsx
@@ -1,43 +1,45 @@
-'use client';
+import { Metadata } from "next";
+import CompaniesPageClient from "./CompaniesPageClient";
 
-import { useState } from 'react';
-import Navbar from '@/components/Navbar';
-import CompanySelector from '@/components/CompanySelector';
-import CompanyQuestionsList from '@/components/CompanyQuestionsList';
-import CompanyProgressSummary from '@/components/CompanyProgressSummary';
-import { type Company } from '@/data/companyQuestions';
+export const metadata: Metadata = {
+  title: "Company-Specific DSA Questions - FAANG & Top Tech Companies",
+  description: "Practice company-specific Data Structures and Algorithms questions from FAANG, Microsoft, Google, Amazon, Meta, and other top tech companies. Curated questions for coding interviews at major tech companies.",
+  keywords: [
+    "company specific DSA questions",
+    "FAANG interview questions",
+    "Google coding interview",
+    "Amazon coding interview",
+    "Microsoft coding interview",
+    "Meta coding interview",
+    "Apple coding interview",
+    "Netflix coding interview",
+    "tech company interview preparation",
+    "coding interview questions by company",
+    "company wise DSA problems"
+  ],
+  openGraph: {
+    title: "Company-Specific DSA Questions - FAANG & Top Tech Companies | DSAMate v2",
+    description: "Practice company-specific Data Structures and Algorithms questions from FAANG, Microsoft, Google, Amazon, Meta, and other top tech companies.",
+    images: [
+      {
+        url: "/og-image.svg",
+        width: 1200,
+        height: 630,
+        alt: "Company-Specific DSA Questions - DSAMate v2",
+      },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Company-Specific DSA Questions - FAANG & Top Tech Companies",
+    description: "Practice company-specific DSA questions from FAANG, Microsoft, Google, Amazon, Meta, and other top tech companies.",
+    images: ["/og-image.svg"],
+  },
+  alternates: {
+    canonical: "/companies",
+  },
+};
 
 export default function CompaniesPage() {
-  const [selectedCompany, setSelectedCompany] = useState<Company | null>(null);
-
-  const handleCompanySelect = (company: Company) => {
-    setSelectedCompany(company);
-  };
-
-  const handleBackToCompanies = () => {
-    setSelectedCompany(null);
-  };
-
-  return (
-    <>
-      <Navbar />
-      <main className="min-h-screen bg-white dark:bg-background text-gray-900 dark:text-white px-4 md:px-12 py-24 transition-colors duration-300">
-        <CompanyProgressSummary />
-        {selectedCompany ? (
-          <>
-            <CompanySelector 
-              onCompanySelect={handleBackToCompanies} 
-              selectedCompany={selectedCompany}
-            />
-            <CompanyQuestionsList 
-              questions={selectedCompany.questions}
-              companyName={selectedCompany.name}
-            />
-          </>
-        ) : (
-          <CompanySelector onCompanySelect={handleCompanySelect} />
-        )}
-      </main>
-    </>
-  );
+  return <CompaniesPageClient />;
 }
